+++ tests/Model/YamlEmptyLineAtEnd/YamlEmptyLineAtEndCheckerTest.php
@@ -6,6 +6,7 @@ namespace YamlStandards\Model\YamlEmptyLineAtEnd;
 
 use PHPUnit\Framework\TestCase;
 use YamlStandards\Model\Config\StandardParametersData;
+use YamlStandards\Model\Config\YamlStandardConfigDefinition;
 use YamlStandards\Result\Result;
 
 /**
@@ -25,7 +26,7 @@ class YamlEmptyLineAtEndCheckerTest extends TestCase
         $yamlEmptyLineAtEndChecker = new YamlEmptyLineAtEndChecker();
 
         foreach ($pathToFiles as $pathToFile) {
-            $standardParametersData = new StandardParametersData(null, null, null, null);
+            $standardParametersData = $this->getStandardsParametersData();
             $result = $yamlEmptyLineAtEndChecker->check($pathToFile, $standardParametersData);
 
             $this->assertSame(Result::RESULT_CODE_INVALID_FILE_SYNTAX, $result->getResultCode());
@@ -44,10 +45,18 @@ class YamlEmptyLineAtEndCheckerTest extends TestCase
         $yamlEmptyLineAtEndChecker = new YamlEmptyLineAtEndChecker();
 
         foreach ($pathToFiles as $pathToFile) {
-            $standardParametersData = new StandardParametersData(null, null, null, null);
+            $standardParametersData = $this->getStandardsParametersData();
             $result = $yamlEmptyLineAtEndChecker->check($pathToFile, $standardParametersData);
 
             $this->assertSame(Result::RESULT_CODE_OK, $result->getResultCode(), sprintf('YAML empty line at end check of "%s" failed.', $pathToFile));
         }
+    }
+
+    /**
+     * @return \YamlStandards\Model\Config\StandardParametersData
+     */
+    private function getStandardsParametersData(): StandardParametersData
+    {
+        return new StandardParametersData(4, 4, 2, YamlStandardConfigDefinition::CONFIG_PARAMETERS_SERVICE_ALIASING_TYPE_VALUE_SHORT);
     }
 }