+++ tests/Model/YamlIndent/YamlIndentCheckerTest.php
@@ -6,6 +6,7 @@ namespace YamlStandards\Model\YamlIndent;
 
 use PHPUnit\Framework\TestCase;
 use YamlStandards\Model\Config\StandardParametersData;
+use YamlStandards\Model\Config\YamlStandardConfigDefinition;
 use YamlStandards\Result\Result;
 
 class YamlIndentCheckerTest extends TestCase
@@ -79,6 +80,6 @@ class YamlIndentCheckerTest extends TestCase
      */
     private function getStandardsParametersData(): StandardParametersData
     {
-        return new StandardParametersData(4, 4, 4, null);
+        return new StandardParametersData(4, 4, 4, YamlStandardConfigDefinition::CONFIG_PARAMETERS_SERVICE_ALIASING_TYPE_VALUE_SHORT);
     }
 }