+++ tests/Model/YamlServiceAliasing/YamlServiceAliasingCheckerTest.php
@@ -83,6 +83,6 @@ class YamlServiceAliasingCheckerTest extends TestCase
      */
     private function getStandardsParametersData(string $type): StandardParametersData
     {
-        return new StandardParametersData(null, null, null, $type);
+        return new StandardParametersData(4, 4, 2, $type);
     }
 }