+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Changed
 - remove command options and move it to config file
+- config: all parameters are now optional and have default value
 
 ### Removed
 - remove support for php 5.6 version and php 7.0 version